// projects/adminmodule/src/app/models/responsemodel.ts
export interface StudentProfileList {
    Aadhaar: number
    FirstName: string
    LastName: string
    FatherName: string
    MotherName: string
    Gender: string
    Dob: string
    CorrepondenceFlatNo: string
    CorrepondenceColonyName: string
    CorrepondenceVillageName: string
    CorrepondenceLandmark: string
    CorrepondenceLocationArea: string
    CorrepondenceCountry: string
    CorrepondenceState: string
    CorrepondenceDistrict: string
    CorrepondenceTaluka: string
    CorrepondenceCity: string
    CorrepondencePincode: number
    Finyear: number
    PhotoFileName: string
    AadhaarFilename: string
    SignatureFileName: string
}